Refactor and split `Ejemplo_Builder.scala` into `DomainExampleApp`
Extracted the main application logic into `DomainExampleApp` to improve code maintainability and separation of concerns. Removed the unused package object and related comments. This refactor simplifies the structure and improves clarity in the codebase.
